chore(lerna): relocate commutable to packages
Kicking off the first of the packages I'm hoping we can push out that will use lerna.
