+++ test/renderer/components/cell/cell-spec.js
@@ -15,7 +15,7 @@ import { Cell } from '../../../../src/notebook/components/cell/cell';
 import {
   emptyCodeCell,
   emptyMarkdownCell,
-} from '../../../../src/commutable';
+} from '../../../../src/packages/commutable';
 
 import { displayOrder, transforms } from '../../../../src/notebook/components/transforms';
