+++ test/renderer/components/cell/code-cell-spec.js
@@ -9,7 +9,7 @@ import CodeCell from '../../../../src/notebook/components/cell/code-cell';
 import {
   emptyCodeCell,
   emptyMarkdownCell,
-} from '../../../../src/commutable';
+} from '../../../../src/packages/commutable';
 
 import { displayOrder, transforms } from '../../../../src/notebook/components/transforms';
