+++ test/renderer/views/draggable-cell-spec.js
@@ -7,7 +7,7 @@ import DraggableCell from '../../../src/notebook/views/draggable-cell';
 
 import {
   emptyMarkdownCell,
-} from '../../../src/commutable';
+} from '../../../src/packages/commutable';
 
 import { displayOrder, transforms } from '../../../src/notebook/components/transforms';
