+++ test/utils.js
@@ -4,11 +4,11 @@ import Immutable from 'immutable';
 import {
   monocellNotebook,
   emptyCodeCell,
-} from '../src/commutable';
+} from '../src/packages/commutable';
 
 import {
   appendCellToNotebook,
-} from '../src/commutable/structures';
+} from '../src/packages/commutable/structures';
 
 import { shutdownKernel } from '../src/notebook/kernel/shutdown';
 import * as actions from '../src/notebook/actions';
